[image][Android] Improved the stability of the memory cache key generation
