+++ packages/expo-image/android/src/main/java/expo/modules/image/ExpoImageViewWrapper.kt
@@ -18,7 +18,6 @@ import com.bumptech.glide.RequestManager
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.bumptech.glide.load.resource.bitmap.DownsampleStrategy
 import com.bumptech.glide.request.RequestOptions
-import com.bumptech.glide.request.target.Target.SIZE_ORIGINAL
 import com.facebook.yoga.YogaConstants
 import expo.modules.image.enums.ContentFit
 import expo.modules.image.enums.Priority
@@ -38,7 +37,6 @@ import expo.modules.kotlin.tracing.trace
 import expo.modules.kotlin.viewevent.EventDispatcher
 import expo.modules.kotlin.views.ExpoView
 import jp.wasabeef.glide.transformations.BlurTransformation
-import java.lang.Float.max
 import java.lang.ref.WeakReference
 import kotlin.math.abs
 import kotlin.math.min
@@ -509,73 +507,7 @@ class ExpoImageViewWrapper(context: Context, appContext: AppContext) : ExpoView(
       newTarget.hasSource = sourceToLoad != null
 
       val downsampleStrategy = if (allowDownscaling) {
-        object : DownsampleStrategy() {
-          var wasTriggered = false
-          override fun getScaleFactor(
-            sourceWidth: Int,
-            sourceHeight: Int,
-            requestedWidth: Int,
-            requestedHeight: Int
-          ): Float {
-            // The method is invoked twice per asset, but we only need to preserve the original dimensions for the first call.
-            // As Glide uses Android downsampling, it can only adjust dimensions by a factor of two,
-            // and hence two distinct scaling factors are computed to achieve greater accuracy.
-            if (!wasTriggered) {
-              newTarget.sourceWidth = sourceWidth
-              newTarget.sourceHeight = sourceHeight
-              wasTriggered = true
-            }
-
-            // The size of the container is unknown, we don't know what to do, so we just run the default scale.
-            if (requestedWidth == SIZE_ORIGINAL || requestedHeight == SIZE_ORIGINAL) {
-              return 1f
-            }
-
-            val aspectRation = calculateScaleFactor(
-              sourceWidth.toFloat(),
-              sourceHeight.toFloat(),
-              requestedWidth.toFloat(),
-              requestedHeight.toFloat()
-            )
-
-            // We don't want to upscale the image
-            return min(1f, aspectRation)
-          }
-
-          private fun calculateScaleFactor(
-            sourceWidth: Float,
-            sourceHeight: Float,
-            requestedWidth: Float,
-            requestedHeight: Float
-          ): Float = when (contentFit) {
-            ContentFit.Contain -> min(
-              requestedWidth / sourceWidth,
-              requestedHeight / sourceHeight
-            )
-            ContentFit.Cover -> max(
-              requestedWidth / sourceWidth,
-              requestedHeight / sourceHeight
-            )
-            ContentFit.Fill, ContentFit.None -> 1f
-            ContentFit.ScaleDown -> if (requestedWidth < sourceWidth || requestedHeight < sourceHeight) {
-              // The container is smaller than the image — scale it down and behave like `contain`
-              min(
-                requestedWidth / sourceWidth,
-                requestedHeight / sourceHeight
-              )
-            } else {
-              // The container is bigger than the image — don't scale it and behave like `none`
-              1f
-            }
-          }
-
-          override fun getSampleSizeRounding(
-            sourceWidth: Int,
-            sourceHeight: Int,
-            requestedWidth: Int,
-            requestedHeight: Int
-          ) = SampleSizeRounding.QUALITY
-        }
+        ContentFitDownsampleStrategy(newTarget, contentFit)
       } else {
         DownsampleStrategy.NONE
       }
